+++ src/platform/emscripten/ffi.rs
@@ -91,6 +91,11 @@ pub type em_fullscreenchange_callback_func = Option<unsafe extern "C" fn(
     fullscreenChangeEvent: *const EmscriptenFullscreenChangeEvent,
     userData: *mut c_void) -> EM_BOOL>;
 
+pub type em_touch_callback_func = Option<unsafe extern "C" fn(
+    eventType: c_int,
+    touchEvent: *const EmscriptenTouchEvent,
+    userData: *mut c_void) -> EM_BOOL>;
+
 #[repr(C)]
 pub struct EmscriptenFullscreenChangeEvent {
     pub isFullscreen: c_int,
@@ -160,6 +165,45 @@ pub struct EmscriptenMouseEvent {
 fn bindgen_test_layout_EmscriptenMouseEvent() {
     assert_eq!(mem::size_of::<EmscriptenMouseEvent>(), 120usize);
     assert_eq!(mem::align_of::<EmscriptenMouseEvent>(), 8usize);
+}
+
+#[repr(C)]
+#[derive(Debug, Copy, Clone)]
+pub struct EmscriptenTouchPoint {
+    pub identifier: c_long,
+    pub screenX: c_long,
+    pub screenY: c_long,
+    pub clientX: c_long,
+    pub clientY: c_long,
+    pub pageX: c_long,
+    pub pageY: c_long,
+    pub isChanged: c_int,
+    pub onTarget: c_int,
+    pub targetX: c_long,
+    pub targetY: c_long,
+    pub canvasX: c_long,
+    pub canvasY: c_long,
+}
+#[test]
+fn bindgen_test_layout_EmscriptenTouchPoint() {
+    assert_eq!(mem::size_of::<EmscriptenTouchPoint>(), 96usize);
+    assert_eq!(mem::align_of::<EmscriptenTouchPoint>(), 8usize);
+}
+
+#[repr(C)]
+#[derive(Debug, Copy, Clone)]
+pub struct EmscriptenTouchEvent {
+    pub numTouches: c_int,
+    pub ctrlKey: c_int,
+    pub shiftKey: c_int,
+    pub altKey: c_int,
+    pub metaKey: c_int,
+    pub touches: [EmscriptenTouchPoint; 32usize],
+}
+#[test]
+fn bindgen_test_layout_EmscriptenTouchEvent() {
+    assert_eq!(mem::size_of::<EmscriptenTouchEvent>(), 3096usize);
+    assert_eq!(mem::align_of::<EmscriptenTouchEvent>(), 8usize);
 }
 
 #[repr(C)]
@@ -238,4 +282,24 @@ extern "C" {
         func: em_callback_func, fps: c_int, simulate_infinite_loop: EM_BOOL);
 
     pub fn emscripten_cancel_main_loop();
+
+    pub fn emscripten_set_touchstart_callback(
+        target: *const c_char, userData: *mut c_void,
+        useCapture: c_int, callback: em_touch_callback_func)
+        -> EMSCRIPTEN_RESULT;
+
+    pub fn emscripten_set_touchend_callback(
+        target: *const c_char, userData: *mut c_void,
+        useCapture: c_int, callback: em_touch_callback_func)
+        -> EMSCRIPTEN_RESULT;
+
+    pub fn emscripten_set_touchmove_callback(
+        target: *const c_char, userData: *mut c_void,
+        useCapture: c_int, callback: em_touch_callback_func)
+        -> EMSCRIPTEN_RESULT;
+
+    pub fn emscripten_set_touchcancel_callback(
+        target: *const c_char, userData: *mut c_void,
+        useCapture: c_int, callback: em_touch_callback_func)
+        -> EMSCRIPTEN_RESULT;
 }